+++ dots/.config/quickshell/ii/panelFamilies/ImmaterialImpulseFamily.qml
@@ -25,6 +25,8 @@ import qs.modules.ii.overlay
 import qs.modules.ii.verticalBar
 import qs.modules.ii.wallpaperSelector
 import qs.modules.ii.desktopMenu
+import qs.modules.ii.dropShelf
+import qs.modules.ii.screenshotResult
 
 Scope {
     PanelLoader { extraCondition: !Config.options.bar.vertical; component: Bar {} }
@@ -50,4 +52,6 @@ Scope {
     PanelLoader { component: WallpaperSelector {} }
     PanelLoader { component: Settings {} }
     PanelLoader { component: DesktopMenu {} }
+    PanelLoader { component: DropShelfPanel {} }
+    PanelLoader { component: ScreenshotResultPanel {} }
 }